mcf: Fix failure caused by having '@' in the branch name
:Release Notes:
Correct the failure of 'mcf --command update' when there's a '@' in the
branch name.

:Detailed Notes:
Having '@' part of a branch name cause git-rev-parse to fail with the
error:

  fatal: ambiguous argument '@branchname@{upstream}': unknown revision or
  path not in the working tree.

Using 'HEAD' in place of current branch name allows the problem to be
worked around.

:Testing Performed:
Successfully ran 'mcf --command update' with @ in branch.

:QA Notes:

:Issues Addressed:
[GF-58476] mcf update fails when branch name contains @

Open-webOS-DCO-1.0-Signed-off-by: Iyad Qumei <iyad.qumei@lge.com>

Change-Id: I35784a69cbbc5cc2915b6d6e51e6e43f887cea7d
Reviewed-on: https://g2g.palm.com/4403
Reviewed-by: DCO Verification
Reviewed-by: Iyad Qumei <iyad.qumei@lge.com>
Tested-by: Iyad Qumei <iyad.qumei@lge.com>
Reviewed-by: Keith Derrick <keith.derrick@lge.com>
Reviewed-by: Susan Montooth <susan.montooth@lge.com>