local.conf: Set a default for SDKMACHINE
:Release Notes:
Add a default setting of "i686" for SDKMACHINE to local.conf . Allow it
to be overridden in webos-local.conf .

:Detailed Notes:
OE expects SDKMACHINE to be set in local.conf when building an SDK.

:Testing Performed:
Confirmed using:
  $ ./mcf.status
  $ cd BUILD-m14tv
  $ bitbake -e pmloglib | grep '^SDKMACHINE='
that this variable is set to "i686". Set it to "x86_64" in
webos-local.conf and used the same bitbake command to confirm that it
was overridden.

:QA Notes:

:Issues Addressed:
[ES-803]   Implement bdk bitbake target
[BHV-7334] CCC: Create BDK for software development outside of OE build
           environment

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I09db20f7321130cd82e524d9eaf789d169ea9241
Reviewed-on: https://g2g.palm.com/6141
Reviewed-by: DCO Verification
Build: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>