mcf: Add --version option (v2.11.0)
:Release Notes:
Allow the different editions of mcf on each of the branches (and in each
the build-* repos) to be distinguished by a version number.

:Detailed Notes:
Unfortunately, the edition on some branches can't match that on master,
so we need a simple way to name distinguish them when documenting
behavior.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[ES-471] Add --version option to mcf

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I282f3357c6bdbe41e1e48131e9e37acc59e4c80a
Reviewed-on: https://g2g.palm.com/6146
Reviewed-by: DCO Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>