mcf: Disable ICECC if it's not installed (v3.0.3)
:Release Notes:
Override definition of ICECC_DISABLED based on ICECC installation
status.

:Detailed Notes:
The implication of enabling ICECC on a build machine that does not have
icecc package installed, is that the make "j" option will assume the
value set for ICECC_PARALLEL_MAKE, potentially overloading the build
machine. This change prevents this from happening, by automatically
checking the installation of ICECC, and setting ICECC_DISABLED to the
appropriate value.

Override ICECC_DISABLED setting based on the installation status of the
"icecc" package on the build machine.  If it is not installed,
ICECC_DISABLED will be set to "1" and PARALLEL_MAKE will not be set to
ICECC_PARALLEL_MAKE.

:Testing Performed:
Verified that ICECC_DISABLED value is "1" when it's not installed.

:QA Notes:

:Issues Addressed:
[BHV-10148] Force ICECC_DISABLED to "1" if ICECC is not installed
[BHV-10505] CCC: Force ICECC_DISABLED to "1" if ICECC is not installed

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I59edb8f4ef8e0f02a841baebcf381dede7dfa6a0
Reviewed-on: https://g2g.palm.com/6149
Reviewed-by: DCO Verification
Build: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Martin Jansa <Martin.Jansa@lge.com>
Tested-by: Martin Jansa <Martin.Jansa@lge.com>
Reviewed-by: Build Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>