oe-init-build-env.in: Make sure that required variables are exported
:Release Notes:
When using Makefile, the MACHINE variable is set, but not exported
before sourcing oe-init-build-env.in, but now we were exporting it
only when the value was empty.

:Detailed Notes:
The variable was correctly respected by the oe-init-build-env.in which
is sourced to the shell, but then the shell used by bitbake didn't see
the value even with MACHINE in BB_ENV_EXTRAWHITE, so it was using
default "qemux86" value from local-conf.in. Change DISTRO assignment
in local-conf.in to be consistent with MACHINE, it's not common to
change it from shell, but it should work OK.

:Testing Performed:
Only build tested.

:QA Notes:
No change to image.

:Issues Addressed:
[ES-1185] MACHINE shell variable not respected by bitbake

Open-webOS-DCO-1.0-Signed-off-by: Martin Jansa <martin.jansa@lge.com>
Change-Id: I54edaae879386d8428fbc832a5fba72e50b1aec2
Reviewed-on: https://g2g.palm.com/6396
Reviewed-by: DCO Verification
Reviewed-by: Susan Montooth <susan.montooth@lge.com>
Tested-by: Susan Montooth <susan.montooth@lge.com>